task: rename folders
