+++ dataworkspace/dataworkspace/static/js/react/features/your-files-v2/YourFiles.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const YourFiles = ({ config }) => {
+
+    return (
+        <></>
+    );
+};
+
+export default YourFiles;
