+++ dataworkspace/dataworkspace/static/js/react/features/your-files-v2/index.js
@@ -1,10 +1,8 @@
 import React, { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
-import App from './App';
 import YourFiles from './YourFiles';
 
-// Workaround for hot reload "ignored-module" issue when developing locally.
 if (module.hot) {
   module.hot.accept();
 }
