+++ dataworkspace/dataworkspace/static/js/react/features/your-files/components/BigDataMessage/index.jsx
