+++ dataworkspace/dataworkspace/static/js/react/features/your-files/components/DeleteObjectsPopup/index.jsx
