+++ dataworkspace/dataworkspace/static/js/react/features/your-files/components/TeamsPrefixMessage/index.jsx
