+++ dataworkspace/dataworkspace/static/js/react/features/your-files/components/UploadFilesPopup/UploadHeaderRow.jsx
