+++ dataworkspace/dataworkspace/static/js/react/features/your-files/components/UploadFilesPopup/index.jsx
