+++ dataworkspace/dataworkspace/static/js/react/features/your-files/index.js
@@ -1,8 +1,9 @@
 import React, { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
-import YourFiles from './YourFiles';
+import App from './App';
 
+// Workaround for hot reload "ignored-module" issue when developing locally.
 if (module.hot) {
   module.hot.accept();
 }
@@ -12,6 +13,6 @@ const root = createRoot(rootElement);
 
 root.render(
   <StrictMode>
-    <YourFiles config={YOURFILES_CONFIG} />
+    <App config={YOURFILES_CONFIG} />
   </StrictMode>
 );
