+++ dataworkspace/dataworkspace/static/js/react/features/your-files/utils.js
